+++ util-hadoop/src/test/java/com/google/cloud/hadoop/util/HadoopCredentialConfigurationTest.java
@@ -81,6 +81,7 @@ public class HadoopCredentialConfigurationTest {
           put(".proxy.address", null);
           put(".proxy.password", null);
           put(".proxy.username", null);
+          put(".auth.impersonation.service.account", null);
         }
       };